# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MonkeyGraph Agent API",
    "description": "MonkeyGraph 知识图谱系统的 Agent API 接口服务文档",
    "version": "1.0.0",
    "contact": {
      "name": "MonkeyGraph Support",
      "url": "https://github.com/czhmisaka/czh_graph"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:13001",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
    "/api/agent/register": {
      "post": {
        "tags": ["Agent"],
        "summary": "注册新 Agent",
        "description": "注册一个新的 Agent，获取 API Key",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Agent 名称"
                  },
                  "description": {
                    "type": "string",
                    "description": "Agent 描述"
                  },
                  "workspace_id": {
                    "type": "string",
                    "description": "绑定的 Workspace ID"
                  },
                  "permissions": {
                    "type": "object",
                    "description": "权限配置"
                  },
                  "monthly_quota": {
                    "type": "integer",
                    "description": "月度请求配额",
                    "default": 100000
                  }
                }
              },
              "example": {
                "name": "my-agent",
                "description": "My AI Agent",
                "permissions": {
                  "graphs": ["read", "write", "delete"],
                  "nodes": ["read", "write", "delete"],
                  "edges": ["read", "write", "delete"]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent 注册成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "agent": {
                    "id": "agent-id",
                    "name": "my-agent",
                    "api_key": "your-api-key",
                    "permissions": {},
                    "monthly_quota": 100000,
                    "requests_used": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器内部错误"
          }
        }
      }
    },
    "/api/agent/me": {
      "get": {
        "tags": ["Agent"],
        "summary": "获取当前 Agent 信息",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "agent": {}
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          }
        }
      },
      "put": {
        "tags": ["Agent"],
        "summary": "更新 Agent 信息",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      }
    },
    "/api/agent/quota": {
      "get": {
        "tags": ["Agent"],
        "summary": "获取配额使用情况",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "quota": {
                    "monthly_quota": 100000,
                    "requests_used": 8,
                    "remaining": 99992,
                    "reset_at": "2026-04-01T00:00:00.000Z"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/rotate-key": {
      "post": {
        "tags": ["Agent"],
        "summary": "轮换 API Key",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "轮换成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "api_key": "new-api-key",
                  "message": "请妥善保存新的 API Key，它只会显示一次"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/graphs/{graphId}/degrees": {
      "get": {
        "tags": ["Algorithms"],
        "summary": "获取节点度统计",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "degrees": [
                    {
                      "id": "node-id",
                      "label": "节点",
                      "type": "person",
                      "in": 2,
                      "out": 1,
                      "total": 3
                    }
                  ],
                  "totalEdges": 5
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/graphs/{graphId}/path": {
      "get": {
        "tags": ["Algorithms"],
        "summary": "路径查找 (BFS)",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "源节点 ID"
          },
          {
            "name": "target",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "目标节点 ID"
          },
          {
            "name": "max_depth",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "最大深度"
          }
        ],
        "responses": {
          "200": {
            "description": "查找成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "path": [],
                  "length": 4,
                  "found": true
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/graphs/{graphId}/nodes/{nodeId}/neighbors": {
      "get": {
        "tags": ["Algorithms"],
        "summary": "邻居查询",
        "security": [
          {
            "AgentAuth": []
          }
        ],
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["all", "incoming", "outgoing"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "node": {
                    "id": "node-id",
                    "label": "节点"
                  },
                  "neighbors": [],
                  "total": 2
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "使用 'Agent your-api-key' 格式进行认证"
      }
    }
  }
}